Combined fork with parrotpy_tryout_bundle
In this commit I i forked the original parrot.py repository and combined it with parrotpy_tryout_bundle and i made a modification to mode_starter.py so that:

- left_click is triggered via sound_pop
- pressing f11 his triggered via click_palatal

I also lowered the power, percentage and times properties of the threshold dictionaries of the respective commands. I'm still unsure exactly what these properties are but they seem to trigger more easily now.